feat(modal):useTextForm.js 훅 추가 & 수정
